pyproject.toml
[project]
name = "asaaspy"
version = "0.3.0"
description = "Unnoficial python client for Asaas"
authors = [
    {name = "Jonhnatha Trigueiro", email = "joepreludian@gmail.com"},
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.8.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "LGPL 2.1"}

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Documentation = "https://asaaspy.readthedocs.io/pt-br/latest"
Source = "https://github.com/joepreludian/asaaspy"
Tracker = "https://github.com/joepreludian/asaaspy/issues"

[project.optional-dependencies]
docs = [
    "mkdocs-material>=9.5.27",
    "mkdocs-mermaid2-plugin>=1.1.1",
    "markdown>=3.6",
]
dev = [
    "pytest>=8.2.2",
    "pytest-cov>=5.0.0",
    "pytest-vcr>=1.0.2",
    "Faker>=26.0.0",
    "factory-boy>=3.3.0",
    "Flake8-pyproject>=1.2.3",
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.coverage.run]
source = ['src']

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.flake8]
ignore = ['N815']
max-line-length = 88
